// types/thing.schema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ThingSchema {
  name: string;
  slug: string;
  /**
   * The chapter in which the character is introduced, especially when they are named. In some cases (such as Maia's funeral-carer) a character is alluded to in a chapter but we don't yet know who they are. In other cases, a a character is present but hasn't been named, in which case we don't want to show their name ahead of time.
   */
  chapter?: number;
  aliases?: {
    chapter: number;
    name: string;
  }[];
  bits: {
    chapter: number;
    text: string;
  }[];
}